+++ build_files/lib/str/sources/put/str_put_str.c
@@ -13,6 +13,6 @@ int str_put_str(const cstr_t string)
 {
     if (!string)
         return EXIT_ERROR;
-    return ((write(1, string, str_len(string)) != str_len(string)) ?
+    return ((write(1, string, str_len(string)) != (ssize_t)str_len(string)) ?
         (EXIT_ERROR) : (EXIT_SUCCESS));
 }
